+++ Catch_core_datacontainer_tests/car_owner_basic2.hpp
@@ -1983,7 +1983,7 @@ namespace cob2 {
 								}
 								if(header.is_property("wheels")) {
 									if(header.is_type("int32_t")) {
-										std::memcpy(car.m_wheels.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), header.record_size));
+										std::memcpy(car.m_wheels.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), size_t(header.record_size)));
 										serialize_selection.car_wheels = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2044,7 +2044,7 @@ namespace cob2 {
 								}
 								if(header.is_property("resale_value")) {
 									if(header.is_type("float")) {
-										std::memcpy(car.m_resale_value.vptr(), reinterpret_cast<float const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(float), header.record_size));
+										std::memcpy(car.m_resale_value.vptr(), reinterpret_cast<float const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(float), size_t(header.record_size)));
 										serialize_selection.car_resale_value = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2115,7 +2115,7 @@ namespace cob2 {
 								}
 								if(header.is_property("age")) {
 									if(header.is_type("int32_t")) {
-										std::memcpy(person.m_age.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(person.size_used) * sizeof(int32_t), header.record_size));
+										std::memcpy(person.m_age.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(person.size_used) * sizeof(int32_t), size_t(header.record_size)));
 										serialize_selection.person_age = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2189,7 +2189,7 @@ namespace cob2 {
 								}
 								if(header.is_property("owner")) {
 									if(header.is_type("uint8_t")) {
-										std::memcpy(car_ownership.m_owner.vptr(), reinterpret_cast<uint8_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(uint8_t), header.record_size));
+										std::memcpy(car_ownership.m_owner.vptr(), reinterpret_cast<uint8_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(uint8_t), size_t(header.record_size)));
 										serialize_selection.car_ownership_owner = true;
 									}
 									else if(header.is_type("uint16_t")) {
@@ -2218,7 +2218,7 @@ namespace cob2 {
 								}
 								if(header.is_property("ownership_date")) {
 									if(header.is_type("int32_t")) {
-										std::memcpy(car_ownership.m_ownership_date.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), header.record_size));
+										std::memcpy(car_ownership.m_ownership_date.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), size_t(header.record_size)));
 										serialize_selection.car_ownership_ownership_date = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2304,7 +2304,7 @@ namespace cob2 {
 								}
 								if(header.is_property("wheels") && mask.car_wheels) {
 									if(header.is_type("int32_t")) {
-										std::memcpy(car.m_wheels.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), header.record_size));
+										std::memcpy(car.m_wheels.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), size_t(header.record_size)));
 										serialize_selection.car_wheels = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2365,7 +2365,7 @@ namespace cob2 {
 								}
 								if(header.is_property("resale_value") && mask.car_resale_value) {
 									if(header.is_type("float")) {
-										std::memcpy(car.m_resale_value.vptr(), reinterpret_cast<float const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(float), header.record_size));
+										std::memcpy(car.m_resale_value.vptr(), reinterpret_cast<float const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(float), size_t(header.record_size)));
 										serialize_selection.car_resale_value = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2436,7 +2436,7 @@ namespace cob2 {
 								}
 								if(header.is_property("age") && mask.person_age) {
 									if(header.is_type("int32_t")) {
-										std::memcpy(person.m_age.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(person.size_used) * sizeof(int32_t), header.record_size));
+										std::memcpy(person.m_age.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(person.size_used) * sizeof(int32_t), size_t(header.record_size)));
 										serialize_selection.person_age = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2510,7 +2510,7 @@ namespace cob2 {
 								}
 								if(header.is_property("owner") && mask.car_ownership_owner) {
 									if(header.is_type("uint8_t")) {
-										std::memcpy(car_ownership.m_owner.vptr(), reinterpret_cast<uint8_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(uint8_t), header.record_size));
+										std::memcpy(car_ownership.m_owner.vptr(), reinterpret_cast<uint8_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(uint8_t), size_t(header.record_size)));
 										serialize_selection.car_ownership_owner = true;
 									}
 									else if(header.is_type("uint16_t")) {
@@ -2539,7 +2539,7 @@ namespace cob2 {
 								}
 								if(header.is_property("ownership_date") && mask.car_ownership_ownership_date) {
 									if(header.is_type("int32_t")) {
-										std::memcpy(car_ownership.m_ownership_date.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), header.record_size));
+										std::memcpy(car_ownership.m_ownership_date.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), size_t(header.record_size)));
 										serialize_selection.car_ownership_ownership_date = true;
 									}
 									else if(header.is_type("int8_t")) {
