+++ Catch_core_datacontainer_tests/car_owner_basic3.hpp
@@ -1935,7 +1935,7 @@ namespace cob3 {
 								}
 								if(header.is_property("_index")) {
 									if(header.is_type("uint16_t")) {
-										std::memcpy(car.m__index.vptr(), reinterpret_cast<uint16_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(uint16_t), header.record_size));
+										std::memcpy(car.m__index.vptr(), reinterpret_cast<uint16_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(uint16_t), size_t(header.record_size)));
 										serialize_selection.car__index = true;
 									}
 									else if(header.is_type("uint8_t")) {
@@ -1964,7 +1964,7 @@ namespace cob3 {
 								}
 								if(header.is_property("wheels")) {
 									if(header.is_type("int32_t")) {
-										std::memcpy(car.m_wheels.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), header.record_size));
+										std::memcpy(car.m_wheels.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), size_t(header.record_size)));
 										serialize_selection.car_wheels = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2025,7 +2025,7 @@ namespace cob3 {
 								}
 								if(header.is_property("resale_value")) {
 									if(header.is_type("float")) {
-										std::memcpy(car.m_resale_value.vptr(), reinterpret_cast<float const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(float), header.record_size));
+										std::memcpy(car.m_resale_value.vptr(), reinterpret_cast<float const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(float), size_t(header.record_size)));
 										serialize_selection.car_resale_value = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2096,7 +2096,7 @@ namespace cob3 {
 								}
 								if(header.is_property("_index")) {
 									if(header.is_type("uint8_t")) {
-										std::memcpy(person.m__index.vptr(), reinterpret_cast<uint8_t const*>(input_buffer), std::min(size_t(person.size_used) * sizeof(uint8_t), header.record_size));
+										std::memcpy(person.m__index.vptr(), reinterpret_cast<uint8_t const*>(input_buffer), std::min(size_t(person.size_used) * sizeof(uint8_t), size_t(header.record_size)));
 										serialize_selection.person__index = true;
 									}
 									else if(header.is_type("uint16_t")) {
@@ -2125,7 +2125,7 @@ namespace cob3 {
 								}
 								if(header.is_property("age")) {
 									if(header.is_type("int32_t")) {
-										std::memcpy(person.m_age.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(person.size_used) * sizeof(int32_t), header.record_size));
+										std::memcpy(person.m_age.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(person.size_used) * sizeof(int32_t), size_t(header.record_size)));
 										serialize_selection.person_age = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2199,7 +2199,7 @@ namespace cob3 {
 								}
 								if(header.is_property("owner")) {
 									if(header.is_type("uint8_t")) {
-										std::memcpy(car_ownership.m_owner.vptr(), reinterpret_cast<uint8_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(uint8_t), header.record_size));
+										std::memcpy(car_ownership.m_owner.vptr(), reinterpret_cast<uint8_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(uint8_t), size_t(header.record_size)));
 										serialize_selection.car_ownership_owner = true;
 									}
 									else if(header.is_type("uint16_t")) {
@@ -2228,7 +2228,7 @@ namespace cob3 {
 								}
 								if(header.is_property("ownership_date")) {
 									if(header.is_type("int32_t")) {
-										std::memcpy(car_ownership.m_ownership_date.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), header.record_size));
+										std::memcpy(car_ownership.m_ownership_date.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), size_t(header.record_size)));
 										serialize_selection.car_ownership_ownership_date = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2314,7 +2314,7 @@ namespace cob3 {
 								}
 								if(header.is_property("_index") && mask.car__index) {
 									if(header.is_type("uint16_t")) {
-										std::memcpy(car.m__index.vptr(), reinterpret_cast<uint16_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(uint16_t), header.record_size));
+										std::memcpy(car.m__index.vptr(), reinterpret_cast<uint16_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(uint16_t), size_t(header.record_size)));
 										serialize_selection.car__index = true;
 									}
 									else if(header.is_type("uint8_t")) {
@@ -2343,7 +2343,7 @@ namespace cob3 {
 								}
 								if(header.is_property("wheels") && mask.car_wheels) {
 									if(header.is_type("int32_t")) {
-										std::memcpy(car.m_wheels.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), header.record_size));
+										std::memcpy(car.m_wheels.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), size_t(header.record_size)));
 										serialize_selection.car_wheels = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2404,7 +2404,7 @@ namespace cob3 {
 								}
 								if(header.is_property("resale_value") && mask.car_resale_value) {
 									if(header.is_type("float")) {
-										std::memcpy(car.m_resale_value.vptr(), reinterpret_cast<float const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(float), header.record_size));
+										std::memcpy(car.m_resale_value.vptr(), reinterpret_cast<float const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(float), size_t(header.record_size)));
 										serialize_selection.car_resale_value = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2475,7 +2475,7 @@ namespace cob3 {
 								}
 								if(header.is_property("_index") && mask.person__index) {
 									if(header.is_type("uint8_t")) {
-										std::memcpy(person.m__index.vptr(), reinterpret_cast<uint8_t const*>(input_buffer), std::min(size_t(person.size_used) * sizeof(uint8_t), header.record_size));
+										std::memcpy(person.m__index.vptr(), reinterpret_cast<uint8_t const*>(input_buffer), std::min(size_t(person.size_used) * sizeof(uint8_t), size_t(header.record_size)));
 										serialize_selection.person__index = true;
 									}
 									else if(header.is_type("uint16_t")) {
@@ -2504,7 +2504,7 @@ namespace cob3 {
 								}
 								if(header.is_property("age") && mask.person_age) {
 									if(header.is_type("int32_t")) {
-										std::memcpy(person.m_age.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(person.size_used) * sizeof(int32_t), header.record_size));
+										std::memcpy(person.m_age.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(person.size_used) * sizeof(int32_t), size_t(header.record_size)));
 										serialize_selection.person_age = true;
 									}
 									else if(header.is_type("int8_t")) {
@@ -2578,7 +2578,7 @@ namespace cob3 {
 								}
 								if(header.is_property("owner") && mask.car_ownership_owner) {
 									if(header.is_type("uint8_t")) {
-										std::memcpy(car_ownership.m_owner.vptr(), reinterpret_cast<uint8_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(uint8_t), header.record_size));
+										std::memcpy(car_ownership.m_owner.vptr(), reinterpret_cast<uint8_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(uint8_t), size_t(header.record_size)));
 										serialize_selection.car_ownership_owner = true;
 									}
 									else if(header.is_type("uint16_t")) {
@@ -2607,7 +2607,7 @@ namespace cob3 {
 								}
 								if(header.is_property("ownership_date") && mask.car_ownership_ownership_date) {
 									if(header.is_type("int32_t")) {
-										std::memcpy(car_ownership.m_ownership_date.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), header.record_size));
+										std::memcpy(car_ownership.m_ownership_date.vptr(), reinterpret_cast<int32_t const*>(input_buffer), std::min(size_t(car.size_used) * sizeof(int32_t), size_t(header.record_size)));
 										serialize_selection.car_ownership_ownership_date = true;
 									}
 									else if(header.is_type("int8_t")) {
